backend/vla_module/requirements.txt
# Python dependencies for Vision-Language-Action (VLA) module

# Core ROS 2 dependencies
rclpy>=3.0.0
rosgraph>=1.16.0

# OpenAI Whisper API
openai>=1.0.0

# Anthropic Claude API (optional)
anthropic>=0.5.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.21.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio processing
soundfile>=0.12.0
webrtcvad>=2.0.10
resampy>=0.4.2
librosa>=0.10.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Math and data processing
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0

# Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0